do not execute e2e tests on Chrome due to protractor incompatibility
Protractor uses its own version of webdriver manager, which is too old to support Chrome 76.

* https://github.com/angular/webdriver-manager/issues/404
* https://github.com/angular/protractor/issues/5289

Its really not worth the effort implementing any of the suggested workarounds.
For now, the e2e tests on Chrome will simply not be executed on Travis.
Once Protractor 5 is updated with the latest webdriver manager version, we can turn them back on again.